create open api parser util
