+++ src/app/services/actions/autocomplete-action-creators.ts
@@ -1,5 +1,6 @@
 import { IAction } from '../../../types/action';
 import { IRequestOptions } from '../../../types/request';
+import { parseOpenApiResponse } from '../../utils/open-api-parser';
 import { AUTOCOMPLETE_FETCH_ERROR, AUTOCOMPLETE_FETCH_PENDING, AUTOCOMPLETE_FETCH_SUCCESS } from '../redux-constants';
 
 export function fetchAutocompleteSuccess(response: object): IAction {
@@ -41,14 +42,13 @@ export function fetchAutoCompleteOptions(url: string): Function {
       const response = await fetch(permissionsUrl, options);
       if (response.ok) {
         const autoCompleteOptions = await response.json();
-        const params = {
+        const parameters = {
           options: autoCompleteOptions,
           url,
           verb: sampleQuery.selectedVerb.toLowerCase()
         };
-        const reduced = getReducedVersion(params);
-
-        return dispatch(fetchAutocompleteSuccess(reduced));
+        const parsedResponse = parseOpenApiResponse(parameters);
+        return dispatch(fetchAutocompleteSuccess(parsedResponse));
       }
 
       throw (response);
@@ -58,34 +58,3 @@ export function fetchAutoCompleteOptions(url: string): Function {
     }
   };
 }
-function getReducedVersion(params: any) {
-  const { options, url, verb } = params;
-  const { paths } = options;
-  try {
-    const parameters: any[] = [];
-    let rootPath = url;
-    if (url.includes('me/drive')) {
-      rootPath = url.substring(3);
-    }
-    const root = paths[`/${rootPath}`];
-    const verbContent = root[`${verb}`];
-    const queryParams = verbContent.parameters;
-    if (queryParams.length > 0) {
-      queryParams.forEach((param: any) => {
-        if (param.name) {
-          const newLocal = {
-            name: param.name,
-            items: param.items.enum || null
-          };
-          parameters.push(newLocal);
-        }
-      });
-    }
-    const resp = { url, parameters, verb };
-    return resp;
-
-  } catch (error) {
-    return { error };
-  }
-}
-
